pyproject.toml
[project]
name = "labububot"
version = "0.1.0"
description = "Bot to get labubus"
authors = [
    {name = "Aareon Sullivan",email = "askully13@gmail.com"}
]
license = {text = "LGPL"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "selenium (>=4.33.0,<5.0.0)",
    "webdriver-manager (>=4.0.2,<5.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "asyncio-mqtt (>=0.16.2,<0.17.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
